--- frontend/app/_api/addMovieList
@@ -1,24 +1,0 @@
-const BASE_URL: string = "http://localhost:3001/listRoutes";
-
-export const createMovieList = async (
-    name: string,
-    userId: string,
-    data: Record<string, any>
-): Promise<any> => {
-    try {
-        const response = await fetch(`${BASE_URL}/addMovieToList`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ name, userId, ...data }),
-        });
-        if (!response.ok) {
-          throw new Error("Internal server error");
-        }
-        return await response.json();
-      } catch (error: any) {
-        console.error("Error adding movie to list:", error.message);
-        throw error;
-      }
-}
