+++ frontend/app/_api/lists.ts
@@ -1,11 +1,10 @@
-const BASE_URL: string = process.env.NEXT_PUBLIC_BACKEND_URL + "/listRoutes";
-const BASE_URL_2: string = process.env.NEXT_PUBLIC_BACKEND_URL + "/movieRoutes";
-const BASE_URL_3: string =
-  process.env.NEXT_PUBLIC_BACKEND_URL + "/authentication";
+const BASE_URL: string = "http://localhost:3001/listRoutes";
+const BASE_URL_2: string = "http://localhost:3001/movieRoutes";
+const BASE_URL_3: string = "http://localhost:3001/authentication";
 
 /**
  * returnsa json object with all list_ids for a userId
- * @param userId
+ * @param userId 
  * @returns json object with all list_ids for a userId
  */
 export const getUserLists = async (userId: string): Promise<any> => {
@@ -31,7 +30,7 @@ export const getUserLists = async (userId: string): Promise<any> => {
  * where entries is an array of objects with fields itemType and item_id,
  * itemType can be 'Movie', 'TV Show'
  * item_id is the id of the movie or TV show
- * @param listId
+ * @param listId 
  * @returns list object
  */
 export const getListInfo = async (listId: string): Promise<any> => {
@@ -52,6 +51,7 @@ export const getListInfo = async (listId: string): Promise<any> => {
   }
 };
 
+
 // error check for movie that is not in db
 export const getMovieInfo = async (movieId: string): Promise<any> => {
   try {
@@ -71,20 +71,14 @@ export const getMovieInfo = async (movieId: string): Promise<any> => {
   }
 };
 
-export const addMovieToList = async (
-  listId: string,
-  movieId: string,
-): Promise<any> => {
+export const addMovieToList = async (listId: string, movieId: string): Promise<any> => {
   try {
-    const response = await fetch(
-      `${BASE_URL}/addMovieToList?listId=${listId}&movieId=${movieId}`,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
+    const response = await fetch(`${BASE_URL}/addMovieToList?listId=${listId}&movieId=${movieId}`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
       },
-    );
+    });
     if (!response.ok) {
       throw new Error("Internal server error");
     }
@@ -95,20 +89,15 @@ export const addMovieToList = async (
   }
 };
 
-export const addList = async (
-  listName: string,
-  userId: string,
-): Promise<any> => {
+
+export const addList = async (listName: string, userId: string): Promise<any> => {
   try {
-    const response = await fetch(
-      `${BASE_URL}/addList?name=${listName}&userId=${userId}`,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
+    const response = await fetch(`${BASE_URL}/addList?name=${listName}&userId=${userId}`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
       },
-    );
+    });
     if (!response.ok) {
       throw new Error("Internal server error");
     }
