+++ frontend/app/_api/profile.ts
@@ -3,7 +3,6 @@ import axios from 'axios';
 const BASE_URL: string = "http://localhost:3001/authentication";
 
 export const getProfileFromToken = async (token: string): Promise<any> => {
-
   try {
     const response = await axios.get(`${BASE_URL}/profile`, {
       headers: {
