+++ frontend/app/_ui/components/EditProfile/EditProfile.tsx
@@ -1,3 +1,4 @@
+// export default EditProfileModal;
 import React, { useState } from "react";
 import ReactModal from "react-modal";
 import { updateUser } from "@/_api/editprofile";
@@ -6,24 +7,26 @@ import styles from "./editprofile.module.css";
 interface EditProfileModalProps {
   isOpen: boolean;
   onClose: () => void;
-  userId: string; // Assuming you have access to the user's ID
+  userId: any; // Assuming you have access to the user's ID
+  user: any;
+  refreshUserData: any;
 }
 
 const EditProfileModal: React.FC<EditProfileModalProps> = ({
   isOpen,
   onClose,
   userId,
+  user,
+  refreshUserData,
 }) => {
   const [formData, setFormData] = useState({
-    email: "",
-    username: "",
-    bio: "",
+    email: user?.email || "",
+    username: user?.username || "",
+    bio: user?.bio || "",
     // Add more fields as needed
   });
 
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
     const { name, value } = e.target;
     setFormData({
       ...formData,
@@ -37,14 +40,14 @@ const EditProfileModal: React.FC<EditProfileModalProps> = ({
       const response = await updateUser(userId, formData);
       console.log("Profile updated:", response);
       onClose();
+      refreshUserData(); 
     } catch (error: any) {
       console.error("Error updating profile:", error.message);
     }
   };
 
   return (
-    <ReactModal isOpen={isOpen} onRequestClose={onClose} ariaHideApp={false}>
-      <div className={styles.modalOverlay}>
+    <ReactModal isOpen={isOpen} onRequestClose={onClose} ariaHideApp={false} className={styles.modalOverlay}>
         <div className={styles.formContainer}>
           <form onSubmit={handleSubmit}>
             <label htmlFor="email">Email:</label>
@@ -54,6 +57,7 @@ const EditProfileModal: React.FC<EditProfileModalProps> = ({
               name="email"
               value={formData.email}
               onChange={handleChange}
+              
               className={styles.formInput}
             />
 
@@ -77,20 +81,11 @@ const EditProfileModal: React.FC<EditProfileModalProps> = ({
             />
 
             <div className={styles.loginBottomButtons}>
-              <button type="submit" className={styles.createAccountButton}>
-                Save Changes
-              </button>
-              <button
-                type="button"
-                onClick={onClose}
-                className={styles.createAccountButton}
-              >
-                Cancel
-              </button>
+              <button type="submit" className={styles.editProfileButton}>Save Changes</button>
+              <button type="button" onClick={onClose} className={styles.editProfileButton}>Cancel</button>
             </div>
           </form>
         </div>
-      </div>
     </ReactModal>
   );
 };
