+++ frontend/app/userpage/page.tsx
@@ -3,30 +3,28 @@ import styles from "./userpage.module.css";
 import React, { useEffect } from "react";
 import { useState } from "react";
 import Image from "next/image";
-import { Tabs, Tab, Box } from "@mui/material";
+import { Tabs, Tab, Box} from '@mui/material';
+import { experimentalStyled as styled } from "@mui/material/styles";
+import Paper from "@mui/material/Paper";
 import profilePic from "@/_assets/sample_profile_pic.png";
+import bannerPic from "@/_assets/sample_banner_pic.jpg";
 import EditProfileModal from "@/_ui/components/EditProfile/EditProfile";
 import { getUserLists, getMovieInfo, addList } from "@/_api/lists";
-import isAuth from "@/protected/protectedRoute";
+import { getUser } from "@/_api/editprofile";
+import notfound from "@/_assets/NOTFOUND.png";
 import { getProfileFromToken } from "@/_api/profile";
-
-/*************************************************************************************************/
-
-import bannerPic from "@/_assets/sample_banner_pic.jpg";
+import isAuth from "@/protected/protectedRoute";
 import SpeedDial from '@mui/material/SpeedDial';
 import SpeedDialIcon from '@mui/material/SpeedDialIcon';
 import SpeedDialAction from '@mui/material/SpeedDialAction';
 import CreateIcon from '@mui/icons-material/Create';
 import EditIcon from '@mui/icons-material/Edit';
 
-
-/*************************************************************************************************/
-
 interface User {
   _id: string;
   username: string;
   email: string;
-  profilePath: string;
+  profilepath: string;
   bio: String;
 }
 
@@ -77,18 +75,17 @@ const Userpage = () => {
   const [isEditProfileOpen, setIsEditProfileOpen] = useState(false);
   const [userLists, setUserLists] = useState<MovieList[]>([]);
   const [isCreateListFormVisible, setIsCreateListFormVisible] = useState(false);
-  const [newListName, setNewListName] = useState("");
+  const [newListName, setNewListName] = useState('');
 
-
+  // this basically is just letting computer know we are in a browser window
   if (typeof window !== "undefined") {
-    useEffect(() => {
-        console.log("here")
+    useEffect(() => {    
         const tokenData = localStorage.getItem("token");
         if (tokenData) {
           // This allows us to parse the token in a usable
           const tokenObject = JSON.parse(tokenData);
           // Hit our profile route 
-          getProfileFromToken(tokenObject.token)
+          getProfileFromToken(tokenObject.value.token)
           // this allows us to unpack the promise we get from the profile route
                 .then(response => {
                   // console.log(response.user.id);
@@ -102,6 +99,7 @@ const Userpage = () => {
       }
     }, []);
   }
+
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     setValue(newValue);
@@ -121,24 +119,27 @@ const Userpage = () => {
 
   const handleCancelClick = () => {
     setIsCreateListFormVisible(false);
-    setNewListName("");
+    setNewListName('');
   };
 
-  // useEffect(() => {
-  //   //const userI;
-  //   fetchUser(userId); // Fetch user data on mount
-  //   fetchUserListsData(userId);
-  // }, []);
+  const actions = [
+    { icon: <CreateIcon />, name: 'Create List', onClick: handleCreateListClick },
+    // { icon: <EditIcon />, name: 'Edit List', onClick: handleEditListClick }, // You'll implement this later
+  ];
+  
+  const fetchUser = async (userId: string) => {
+    try {
+      const data = await getUser(userId); 
+      setUser(data);
+    } catch (error) {
+      console.error("Error fetching user data", error);
+    }
+  };
 
-  const fetchUser = (userId: string) => {
-    // Make API call to fetch user data
-    fetch(
-      process.env.NEXT_PUBLIC_BACKEND_URL +
-        `/authentication/getUser?userId=${userId}`,
-    )
-      .then((response) => response.json())
-      .then((data) => setUser(data))
-      .catch((error) => console.error("Error fetching user data", error));
+  const refreshUserData = () => {
+    if (user){
+      fetchUser(user._id);
+    }
   };
   
 
@@ -148,37 +149,47 @@ const Userpage = () => {
       for (let list of lists) {
         for (let entry of list.entries) {
           const movieInfo = await getMovieInfo(entry.item_id);
-          entry.imageUrl = `https://image.tmdb.org/t/p/original${movieInfo.image_path}`;
+          entry.imageUrl = `https://image.tmdb.org/t/p/original${movieInfo.image_path}`
         }
       }
-
+      
       setUserLists(lists);
+
     } catch (error) {
       console.error("Error fetching user lists", error);
     }
   };
 
-  const handleCreateListSubmit = async (
-    event: React.FormEvent<HTMLFormElement>,
-  ) => {
+
+  const handleCreateListSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     if (user) {
       const newList = await addList(newListName, user._id); // Replace with your API function and user ID
       setUserLists([...userLists, newList]);
+      refreshUserData();
     }
     setIsCreateListFormVisible(false);
-    setNewListName("");
+    setNewListName('');
   };
-
+  
+ 
   return (
     <div className={styles.userPage}>
+      <div className={styles.banner}>
+        <Image
+          src= {bannerPic} // src={user?.bannerPic || defaultBannerPic}
+          layout="fill"
+          objectFit="cover"
+          alt="Banner Pictrue"
+        />
+      </div>
       <div className={styles.userInfo}>
         <div className={styles.photoUsername}>
           <Image
             priority
             src={profilePic} // src={user?.profilePic || profilePic}
-            width={500}
-            height={500}
+            width={200}
+            height={200}
             alt="Profile Picture"
             className={styles.profilePicture}
           />
@@ -205,60 +216,74 @@ const Userpage = () => {
             <EditProfileModal
               isOpen={isEditProfileOpen}
               onClose={closeEditProfileModal}
-              userId={''}
+              userId={user?._id}
+              user={user}
+              refreshUserData={refreshUserData}
             />
           )}
         </div>
       </div>
-      <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
+      <Box>
         <Tabs
           value={value}
           onChange={handleChange}
           centered
-          TabIndicatorProps={{ style: { backgroundColor: "blue" } }}
-          sx={{ borderBottom: 1, borderColor: "white" }}
+          textColor="inherit"
+          TabIndicatorProps={{ style: { backgroundColor: "white" } }}
+          sx={{ borderBottom: 0.5, borderColor: "white" }}
         >
           <Tab
             label="Movie Lists"
             {...a11yProps(0)}
             sx={{
-              color: value === 0 ? "blue" : "white", // Set the text color based on the tab's selection
+              color: value === 0 ? "white" : "white", // Set the text color based on the tab's selection
             }}
           />
           <Tab
             label="Watchlist"
             {...a11yProps(1)}
             sx={{
-              color: value === 1 ? "blue" : "white",
+              color: value === 1 ? "white" : "white",
             }}
           />
           <Tab
             label="Ratings"
             {...a11yProps(2)}
             sx={{
-              color: value === 2 ? "blue" : "white",
+              color: value === 2 ? "white" : "white",
             }}
           />
         </Tabs>
       </Box>
       <CustomTabPanel value={value} index={0}>
-        <div className={styles.MovieLists}>
-          <div className={styles.MovieButtons}>
-            <button className={styles.addMovieList}onClick={handleCreateListClick}>Create List</button>
+        <div className={styles.movieLists}>
+          <div className={styles.movieButtons}>
+            <div>
+              <SpeedDial
+                ariaLabel="SpeedDial openIcon example"
+                direction="right"
+                icon={<SpeedDialIcon />}
+                className={styles.SpeedDial}
+              >
+                {actions.map((action) => (
+                  <SpeedDialAction
+                    key={action.name}
+                    icon={action.icon}
+                    tooltipTitle={action.name}
+                    onClick={action.onClick}
+                  />
+                ))}
+              </SpeedDial>
+              </div>
               {isCreateListFormVisible && (
-                <form
-                  className={styles.addMovieListForm}
-                  onSubmit={handleCreateListSubmit}
-                >
+                <form className={styles.addMovieListForm} onSubmit={handleCreateListSubmit}>
                   <input
                     type="text"
                     value={newListName}
                     onChange={(e) => setNewListName(e.target.value)}
                   />
-                  <button type="submit">Submit</button>
-                  <button type="button" onClick={handleCancelClick}>
-                    Cancel
-                  </button>
+                  <button type="submit">Create</button>
+                  <button type="button" onClick={handleCancelClick}>Cancel</button>
                 </form>
               )}
             </div>
@@ -269,7 +294,7 @@ const Userpage = () => {
                   {list.entries.map((entry, index) => (
                     <div key={index} className={styles.imageItem}>
                       {entry.imageUrl ? (
-                        <Image
+                        <img
                           src={entry.imageUrl}
                           alt={entry.item_id}
                           style={{
@@ -286,13 +311,16 @@ const Userpage = () => {
                 </div>
               </div>
             ))}
-          </div>
-        </CustomTabPanel>
-        <CustomTabPanel value={value} index={1}></CustomTabPanel>
-        <CustomTabPanel value={value} index={2}></CustomTabPanel>
-      </div>
-    )                    
-  }
+        </div>
+      </CustomTabPanel>
+      <CustomTabPanel value={value} index={1}>
+        
+      </CustomTabPanel>
+      <CustomTabPanel value={value} index={2}>
+        
+      </CustomTabPanel>
+    </div>
+  );
+}
 
 export default isAuth(Userpage);
-
